ai/requirements.txt
requests
openai>=1.0.0
python-dotenv